feat: 线程方法优化, 现在支持主线程和工作线程之间的持续通信

+++ packages/fe/worker/worker.ts
@@ -4,7 +4,11 @@
  * @param data 发送到worker的数据
  * @returns
  */
-export function runWorkerOnce<T = any>(filePath: string | URL, data: any) {
+export function runWorkerOnce<T = any>(
+  filePath: string | URL,
+  data: any,
+  cb?: (data: T) => void
+) {
   return new Promise<T>((rs, rj) => {
     const worker = new Worker(filePath)
 
@@ -13,9 +17,18 @@ export function runWorkerOnce<T = any>(filePath: string | URL, data: any) {
       worker.terminate()
     }
 
-    worker.onmessage = function (e) {
-      rs(e.data)
-      worker.terminate()
+    worker.onmessage = function (
+      e: MessageEvent<{
+        event: 'end' | 'data'
+        data: T
+      }>
+    ) {
+      if (e.data.event === 'end') {
+        rs(e.data.data)
+        worker.terminate()
+      } else {
+        cb?.(e.data.data)
+      }
     }
 
     worker.postMessage(data)
@@ -24,7 +37,10 @@ export function runWorkerOnce<T = any>(filePath: string | URL, data: any) {
 
 interface WorkerRunnerOptions<T> {
   onError?: (err: any) => void
-  onMessage: (data: T) => void
+  onMessage: (data: {
+    event: 'end' | 'send';
+    data: T
+  }) => void
 }
 
 class WorkerRunner<T> {
@@ -72,4 +88,31 @@ export function runWorker<T = any>(
   options: WorkerRunnerOptions<T>
 ) {
   return new WorkerRunner(filePath, options)
+}
+
+type IWorker = {
+  send: <D>(data: D) => void
+  end: <D>(data: D) => void
+}
+
+type WorkerFn<T = any> = (accept: T, worker: IWorker) => any
+
+export function createWorker<T>(fn: WorkerFn<T>) {
+  const worker: IWorker = {
+    send(data) {
+      window.postMessage({
+        type: 'data',
+        data
+      })
+    },
+    end(data) {
+      window.postMessage({
+        event: 'end',
+        data
+      })
+    }
+  }
+  window.onmessage = function (e: MessageEvent<T>) {
+    fn(e.data, worker)
+  }
 }
